Extract HomepageCustomizableSection
